workhub-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
pyodbc==5.1.0
pymssql==2.2.11
python-dotenv==1.0.0
marshmallow==3.20.1
cryptography==41.0.7
email-validator==2.1.0
dnspython==2.4.2
pandas==2.1.4
bleach==6.1.0
werkzeug==3.0.1
# GCP Dependencies
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
gunicorn==21.2.0